Caching logger enable state to avoid lenghty checks with every log call
